+++ packages/orq-rc/src/orq_ai_sdk/models/agentstartedstreamingevent.py
@@ -7,7 +7,13 @@ from .filepart import FilePart, FilePartTypedDict
 from .textpart import TextPart, TextPartTypedDict
 from .toolcallpart import ToolCallPart, ToolCallPartTypedDict
 from .toolresultpart import ToolResultPart, ToolResultPartTypedDict
-from orq_ai_sdk.types import BaseModel, UNSET_SENTINEL
+from orq_ai_sdk.types import (
+    BaseModel,
+    Nullable,
+    OptionalNullable,
+    UNSET,
+    UNSET_SENTINEL,
+)
 from orq_ai_sdk.utils import get_discriminator
 import pydantic
 from pydantic import Discriminator, Tag, model_serializer
@@ -344,6 +350,7 @@ class AgentStartedStreamingEventDataTypedDict(TypedDict):
     agent_key: str
     integration_id: NotRequired[str]
     settings: NotRequired[SettingsTypedDict]
+    agent_description: NotRequired[Nullable[str]]
     variables: NotRequired[Dict[str, Any]]
     tool_execution_id: NotRequired[str]
     is_continuation: NotRequired[bool]
@@ -370,6 +377,8 @@ class AgentStartedStreamingEventData(BaseModel):
 
     settings: Optional[Settings] = None
 
+    agent_description: OptionalNullable[str] = UNSET
+
     variables: Optional[Dict[str, Any]] = None
 
     tool_execution_id: Optional[str] = None
@@ -386,6 +395,7 @@ class AgentStartedStreamingEventData(BaseModel):
             [
                 "integration_id",
                 "settings",
+                "agent_description",
                 "variables",
                 "tool_execution_id",
                 "is_continuation",
@@ -393,15 +403,24 @@ class AgentStartedStreamingEventData(BaseModel):
                 "responseId",
             ]
         )
+        nullable_fields = set(["agent_description"])
         serialized = handler(self)
         m = {}
 
         for n, f in type(self).model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
+            is_nullable_and_explicitly_set = (
+                k in nullable_fields
+                and (self.__pydantic_fields_set__.intersection({n}))  # pylint: disable=no-member
+            )
 
             if val != UNSET_SENTINEL:
-                if val is not None or k not in optional_fields:
+                if (
+                    val is not None
+                    or k not in optional_fields
+                    or is_nullable_and_explicitly_set
+                ):
                     m[k] = val
 
         return m