+++ packages/orq-rc/src/orq_ai_sdk/models/createconversationop.py
@@ -199,7 +199,7 @@ class CreateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh3zn2r2farzj0szrjj7rbn7"
+        "conv_01kh5206few1spvqzbfx9hay2c"
     )
     r"""Unique conversation identifier with `conv_` prefix."""