+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3ZN3B0C1VWA6DJR5NAXG8V"
+        "01KH52070YY7ZX7BE3M55E2VK2"
     )
     r"""The unique identifier of the data source"""