+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2093,7 +2093,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KH3ZN2K8Z21D9A2YYPN024MF"
+    id: Optional[str] = "01KH5206AWBTNA9PFAYBKT3FA2"
 
     description: Optional[str] = None