+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH3ZN3B4SVQVFQ4S0HWRJGHY"
+        "01KH520713V3400Z52YQAR51Z7"
     )
     r"""The unique identifier of the data source"""